pyproject.toml
[tool.poetry]
name = "postgres-homeworks-2023"
version = "0.1.0"
description = ""
authors = ["Evgenii Ermak <djermak3000@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
